pyproject.toml
[project]
name = "storyteller"
version = "0.1.0"
description = "AI Storyteller and DM Assistant"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "google-generativeai>=0.3.0",
    "openai>=1.0.0",
    "anthropic>=0.3.0",
    "mcp>=0.1.0",
    "pydantic>=2.0.0"
]

[project.scripts]
storyteller = "storyteller.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
